+++ src/private/WRLDRouteView.m
@@ -82,8 +82,8 @@ static double VERTICAL_LINE_HEIGHT = 5.0;
 {
    WRLDPolyline* polyline = [WRLDPolyline polylineWithCoordinates:step.path count:step.pathCount onIndoorMap:step.indoorId onFloor:floor];
     double elevations[2];
-    elevations[0] = 0;
-    elevations[1] = height;
+    elevations[0] = (CGFloat)0;
+    elevations[1] = (CGFloat)height;
     [polyline setPerPointElevations:elevations count:2];
     return polyline;
 }
